feat: Add PR Size Labeler Github Action
Fixes #5

The PR Size Labeler Github Action has been added to the project, which
would label the future PRs to the project based on the number of lines added,
modified and deleted. The default size categories are used for the configuration,
defining them explicitly to be visible to the repo contributors. The labels were
customized however to be more human readable.